refactor(core): inline the diagnostics catalog into diagnostics.ts
Move diagnosticCatalog and the DiagnosticDoc type out of the separate
diagnosticCatalog.ts and into diagnostics.ts, next to Diagnostics.explain
that uses them. Re-export from index.ts and types.ts unchanged.
