+++ packages/core/src/index.ts
@@ -3,8 +3,7 @@ export * as ast from '@kubb/ast'
 export { diagnosticCode, logLevel } from './constants.ts'
 export { createAdapter } from './createAdapter.ts'
 export { createDebugger } from './createDebugger.ts'
-export { diagnosticCatalog } from './diagnosticCatalog.ts'
-export { DiagnosticError, Diagnostics } from './diagnostics.ts'
+export { diagnosticCatalog, DiagnosticError, Diagnostics } from './diagnostics.ts'
 export { createKubb } from './createKubb.ts'
 export { createRenderer } from './createRenderer.ts'
 export { createStorage } from './createStorage.ts'
